[DDW-321] Separates Nodes requests into their own sub-directory
